Moved common logic to hook and memorized additional row heights
