+++ src/components/org/OrgChart/components/Children.tsx
@@ -3,6 +3,7 @@ import React, { useEffect, useContext, useMemo, useCallback } from 'react';
 import Card from './Card';
 import { OrgChartContext, OrgChartContextReducer } from '../store';
 import { OrgNode } from '../orgChartTypes';
+import useAdditionalRowHeight from '../useAdditionalRowHeight';
 
 function Children<T>() {
     const {
@@ -17,7 +18,6 @@ function Children<T>() {
             numberOfCardsPerRow,
             width,
             additionalChildRowHeight,
-            additionalAsideRowHeight,
         },
         dispatch,
     } = useContext<OrgChartContextReducer<T>>(OrgChartContext);
@@ -46,8 +46,6 @@ function Children<T>() {
         [childrenNodes, numberOfCardsPerRow]
     );
 
-    const initialMargin = (rowMargin - 20) * asideRows + (numberOfCardsPerRow === 1 ? 40 : 50);
-
     useEffect(() => {
         if (rows.length !== childrenRows) {
             dispatch({
@@ -57,7 +55,19 @@ function Children<T>() {
         }
     }, [childrenRows, rows]);
 
-   
+    const additionalRowHeight = useAdditionalRowHeight(rows);
+
+    useEffect(() => {
+        const maxAdditionalRowHeight = additionalRowHeight[additionalRowHeight.length - 1];
+        if (maxAdditionalRowHeight !== additionalChildRowHeight) {
+            dispatch({
+                type: 'UPDATE_ADDITIONAL_ROW_HEIGHT',
+                additionalChildRowHeight: maxAdditionalRowHeight,
+            });
+        }
+    }, [additionalRowHeight]);
+
+    const initialMargin = (rowMargin - 20) * asideRows + (numberOfCardsPerRow === 1 ? 40 : 50);
 
     const getStartXPosition = (cards: OrgNode<T>[], rowNo: number) => {
         if (numberOfCardsPerRow === 1) {
@@ -90,50 +100,6 @@ function Children<T>() {
         },
         [centerX, cardWidth, cardMargin, rowMargin, initialMargin, rows, width]
     );
-
-    const getAdditionalRowHeight = useCallback(
-        (cards: OrgNode<T>[]) => {
-            const greatestAdditionPersons = cards.reduce((highest: number, card: OrgNode<T>) => {
-                if (card.numberOfAssignees && card.numberOfAssignees > highest) {
-                    return card.numberOfAssignees;
-                }
-                return highest;
-            }, 0);
-            if (greatestAdditionPersons === 0) {
-                return 0;
-            }
-            return (
-                Math.ceil(greatestAdditionPersons / 3) *
-                (numberOfCardsPerRow === 1 ? rowMargin - 20 : rowMargin)
-            );
-        },
-        [numberOfCardsPerRow, rowMargin]
-    );
-
-    const additionalRowHeight = rows.reduce(
-        (additionalHeight: number[], _: OrgNode<T>[], index: number) => {
-            const previousHeight = additionalHeight.reduce(
-                (totalHeight: number, height: number) => totalHeight + height,
-                0
-            );
-            if (index === 0) {
-                return [...additionalHeight, additionalAsideRowHeight];
-            }
-            const currentRowHeight = getAdditionalRowHeight(rows[index - 1]);
-            return [...additionalHeight, previousHeight + currentRowHeight];
-        },
-        []
-    );
-
-    useEffect(() => {
-        const maxAdditionalRowHeight = additionalRowHeight[additionalRowHeight.length - 1];
-        if (maxAdditionalRowHeight !== additionalChildRowHeight) {
-            dispatch({
-                type: 'UPDATE_ADDITIONAL_ROW_HEIGHT',
-                additionalChildRowHeight: maxAdditionalRowHeight,
-            });
-        }
-    }, [additionalRowHeight]);
 
     return (
         <g className="children">
